pyproject.toml
[project]
name = "notebooks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ebooklib>=0.20",
    "lingua-language-detector>=2.1.1",
    "marimo>=0.18.1",
    "nbformat>=5.10.4",
    "pyzmq>=27.1.0",
]
